requirements/test.txt
-r ./base.txt

mypy==0.620

pylint==2.1.1
pylint-common==0.2.5
pylint-django==2.0.2

pytest==3.8.0
pytest-sugar==0.9.1
pytest-django==3.4.2
pytest-asyncio==0.9.0

coverage==4.5.1
django-coverage-plugin==1.6.0

flake8==3.5.0
isort==4.3.4
yapf==0.24.0
